+++ validation/kubernetes/control-plane-logs/main.tf
@@ -1,0 +1,16 @@
+terraform {}
+
+provider "kubernetes" {}
+
+provider "helm" {}
+
+module "control-plane-logs" {
+  source = "../../../modules/kubernetes/control-plane-logs"
+
+  providers = {
+    kubernetes = kubernetes
+    helm       = helm
+  }
+
+  cloud_provider   = "azure"
+}